Fix wrong response is tests_view_cluster_activity
The problem in the test was that Starlette Test Client opens a new
connection and start new session, while flask test client
uses the same database session. The test did not show data because
the data was not committed and session was not closed - which also
failed sqlite local tests with "database is locked" error.

This solution can be applied to other tests where data is differnet
than expected and there is a missing commit / close when data is
prepared.